Sorted files
Organized files for better interpretation of the code.
