+++ Languages.py
@@ -1,0 +1,2 @@
+def get_lyrics_language(tracks_with_lyrics):
+    return tracks_with_lyrics
