+++ Lyrics.py
@@ -1,0 +1,2 @@
+def get_tracks_lyrics(tracks):
+    return tracks
